fix: styleguide로 page 작업 옮김
